moving search input and reset button into its own component
