+++ src/app/components/AdvocatesSearch/index.ts
@@ -1,0 +1,2 @@
+export { default as AdvocatesSearch } from './AdvocatesSearch';
+export type { AdvocatesSearchProps } from './types';
