+++ src/app/components/AdvocatesSearch/types.ts
@@ -1,0 +1,3 @@
+export type AdvocatesSearchProps = {
+  searchTermCallback: (searchTerm: string) => void;
+};
